# Week3/Day1/backend/task-manager-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Manager API",
    "version": "1.0.0",
    "description": "A simple Task Manager API with CRUD operations"
  },
  "servers": [
    {
      "url": "https://week3-day1-task-backend-apis.vercel.app",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/tasks": {
      "get": {
        "summary": "Get all tasks",
        "responses": {
          "200": {
            "description": "List of tasks",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    { "id": 1, "title": "Learn Express", "completed": false }
                  ],
                  "message": "Tasks retrieved successfully"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "example": "New task" },
                  "completed": { "type": "boolean", "example": false }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": { "id": 2, "title": "New task", "completed": false },
                  "message": "Task created successfully"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "summary": "Get a task by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": { "id": 1, "title": "Learn Express", "completed": false },
                  "message": "Task retrieved successfully"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "data": null,
                  "message": "Task not found"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "Updated title" },
                  "completed": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": { "id": 1, "title": "Updated title", "completed": true },
                  "message": "Task updated successfully"
                }
              }
            }
          },
          "400": { "description": "Validation error" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": null,
                  "message": "Task deleted successfully"
                }
              }
            }
          },
          "404": { "description": "Task not found" }
        }
      }
    }
  }
}